Moved PoolAllocator from main.cc into separate files.
