+++ constants.hh
@@ -1,0 +1,9 @@
+#pragma once
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+#define BUFFER_SIZE 1024
+#define LARGE_BUFFER_SIZE 1024 * 1024 * 3
+#define DEBUG 0
+
+#endif  // CONSTANTS_H
